fix(config): close the four review gaps left after the alarm-clock fix
Addresses the remaining points from the adversarial pre-merge review on
#2220 (the deferred-removal registration itself already landed as
670c954e):

- CONFIGURATION.md no longer contradicts itself: the env-var table's
  "reserved key stays reserved" note now calls out that
  VELESDB_STORAGE_DATA_DIR is stronger than reserved (deprecated), the
  raw [storage] example annotates each of the three deprecated fields
  instead of restating stale functional claims (the old "64 - 65536 (64
  GB max)" range contradicted KNOWN_LIMITATIONS.md's 1 TiB cap), and
  both Usage Examples recipes drop the dead fields rather than teach
  them as tuning knobs.
- The mobile and tauri-plugin example configs no longer teach
  mmap_cache_mb/vector_alignment as real tuning levers ("far too
  generous for a phone", "matches the cache line") — replaced with a
  deprecation note; values are unchanged.
- The Python surface (options.rs) now documents the same deprecated /
  reserved split as the core doc, appended to the existing per-field
  doc lines rather than added as new ones so the file's frozen
  line-budget baseline (1123) does not move.
- A new test proves warn_deprecated_storage_fields() actually fires a
  tracing::warn! naming the deviating field, not just that the pure
  deprecated_storage_entries() list is right — deleting the warn! call
  from validate() now fails a test instead of shipping silently. The
  warn-capturing subscriber this needs already existed in
  scan_cap_observability_tests.rs; moved into fixtures_tests.rs as a
  shared helper (capture_warns) instead of duplicating it.
- Corrected the "same accept-and-warn cycle as [wal_batch]" claim
  everywhere it appears (config.rs, config_validation.rs,
  CONFIGURATION.md, CHANGELOG.md): [wal_batch] has no validation at
  all, while mmap_cache_mb keeps its pre-existing hard range check —
  left as is on purpose, not loosened as a second change bundled into
  this one.
- Renamed test_still_unwired_sections_are_reported_wholesale to
  test_search_reported_wholesale_storage_mode_reported_by_field: the
  test asserts a single field for storage now, not a section, and the
  old name said the opposite.

Verified: cargo fmt --check (workspace), cargo clippy -p velesdb-core
--all-targets --features persistence,gpu,update-check -D warnings -D
pedantic (clean), PYO3_PYTHON=python3 cargo clippy -p velesdb-python
--lib -D warnings (clean), cargo test -p velesdb-core --features
persistence (storage + scan_cap + the two new/renamed tests, all
green), check-deferred-removals.py, check-file-budgets.py,
check-ai-attribution.py, check-doc-freshness.py (all five guards),
check-doc-contract.sh, check-mcp-doc-contract.py.

Merges origin/develop in (this branch had fallen behind).
